% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlesort.R
\name{titlesort}
\alias{titlesort}
\title{Predicts the seniority of contacts based on their job titles}
\usage{
titlesort(input.title)
}
\arguments{
\item{input.title}{Vector of strings of job titles}
}
\value{
Returns a vector of 1s and 0s corresponding to predicted job titles.
}
\description{
\code{titlesort} returns a vector of 1s and 0s corresponding to the predicted
seniority of input job titles. A prediction of 1 corresponds to a director-level-or-higher
position, and a contact with a 0 is predicted to be an individual contributor
or junior manager. Based on a bag of words method, it has 95%+ precision and recall
for most English-language job titles, with a slight bias towards the technology industry.
Also works OK for German and French titles, and should not be used for other languages.
\code{titlesort} is embedded inside \code{\link{mail.subject}} and does not need to be used
separately with that function.
}
\examples{
titlesort(mail$title)

}
